Remove old app.py file, fix threads at launch
Ref #235 Improve App start, fix Unit Tests for AlignakApp class
