+++ test/test_app.py
@@ -24,7 +24,7 @@ import sys
 import unittest2
 from PyQt5.QtWidgets import QApplication
 
-from alignak_app.app import AlignakApp, AppProgressBar
+from alignak_app.app import AppProgressBar
 
 
 class TestApp(unittest2.TestCase):
@@ -40,20 +40,6 @@ class TestApp(unittest2.TestCase):
             cls.app = QApplication(sys.argv)
         except:
             pass
-
-    # def test_app_reconnect_mode(self):
-    #     """Reconnect App"""
-    #
-    #     under_test = AlignakApp()
-    #
-    #     self.assertIsNone(under_test.tray_icon)
-    #     self.assertFalse(under_test.reconnect_mode)
-    #
-    #     # Build alignak_app
-    #     under_test.app_reconnecting_mode('ERROR')
-    #
-    #     self.assertIsNone(under_test.tray_icon)
-    #     self.assertTrue(under_test.reconnect_mode)
 
     def test_app_progressbar(self):
         """App Progress Bar"""
